Major Bug fixes and feature changes
196 changes and 3+ hours of work. For bug fixes.

When Anonymous Websites was first created, two teenagers who had just learned to make a simple HTML page with a header and paragraph wanted to make a website for their friends.
These teenagers have come very far, learning more HTML, CSS, JavaScript, and PHP.

We really took our time to fix issues and to keep the page modern and updated, as well as learn new and better methods for providing a website. We hope you will continue to enjoy the updated Anonymous Websites.