Large UI Overhaul + More
The largest component of this commit is the UI overhaul. We changed the tutorial pages the most, and kept a more uniform and simple design (much easier on the eyes!).

We also added a location to access news articles when we make them. You can find tech-based information from the navigation bar. We make articles when we have the time, and don't have a set schedule.

We also decided to remove the Experimental Pages section. It has been unused for quite some time, and it wasn't being actively developed.

We also cleaned up the repository a bit and removed unused files.